🚚 Rename: 공지사항, 회원가입을 common으로 이동 (url 변경됨)
