+++ pages/common/notice/noticedetail.module.scss
